requirements.txt
asgiref
attrs
branca
certifi
charset-normalizer
coverage
Django
django-redis
djangorestframework
drf-spectacular
folium
idna
inflection
iniconfig
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
numpy
packaging
pandas
pluggy
Pygments
pytest
pytest-cov
python-dateutil
PyYAML
redis
referencing
requests
rpds-py
scipy
six
sqlparse
typing_extensions
uritemplate
urllib3
xyzservices
